Feedback + Go version gate
