+++ buildspec/linuxIntegrationTests.yml
@@ -25,6 +25,7 @@ phases:
             - '>/dev/null pip3 install --upgrade aws-sam-cli'
             - '>/dev/null pip3 install --upgrade awscli'
             - '>/dev/null pip3 install pylint'
+            # Install Go
             - '>/dev/null VERSION=$(curl "https://golang.org/VERSION?m=text") && wget -q "https://dl.google.com/go/$VERSION.linux-amd64.tar.gz"'
             - '>/dev/null tar -C /usr/local -xzf $VERSION.linux-amd64.tar.gz && export PATH=$PATH:/usr/local/go/bin && export GOPROXY=direct'
             # login to DockerHub so we don't get throttled
